+++ src/test/java/com/snowflake/kafka/connector/streaming/iceberg/IcebergIngestionSchemaEvolutionIT.java
@@ -1,0 +1,100 @@
+package com.snowflake.kafka.connector.streaming.iceberg;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.snowflake.kafka.connector.Utils;
+import com.snowflake.kafka.connector.internal.DescribeTableRow;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import org.apache.kafka.connect.sink.SinkRecord;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+
+public class IcebergIngestionSchemaEvolutionIT extends IcebergIngestionIT {
+
+  @Override
+  protected Boolean isSchemaEvolutionEnabled() {
+    return true;
+  }
+
+  @ParameterizedTest()
+  @MethodSource("prepareData")
+  @Disabled
+  void shouldEvolveSchemaAndInsertRecords(
+      String description, String message, DescribeTableRow[] expectedSchema, boolean withSchema)
+      throws Exception {
+    // start off with just one column
+    List<DescribeTableRow> rows = describeTable(tableName);
+    assertThat(rows.size()).isEqualTo(1);
+    assertThat(rows.get(0).getColumn()).isEqualTo(Utils.TABLE_COLUMN_METADATA);
+
+    SinkRecord record = createKafkaRecord(message, 0, withSchema);
+    service.insert(Collections.singletonList(record));
+    waitForOffset(-1);
+    rows = describeTable(tableName);
+    assertThat(rows.size()).isEqualTo(9);
+
+    // don't check metadata column schema, we have different tests for that
+    rows =
+        rows.stream()
+            .filter(r -> !r.getColumn().equals(Utils.TABLE_COLUMN_METADATA))
+            .collect(Collectors.toList());
+
+    assertThat(rows).containsExactlyInAnyOrder(expectedSchema);
+
+    // resend and store same record without any issues now
+    service.insert(Collections.singletonList(record));
+    waitForOffset(1);
+
+    // and another record with same schema
+    service.insert(Collections.singletonList(createKafkaRecord(message, 1, withSchema)));
+    waitForOffset(2);
+
+    // and another record with extra field - schema evolves again
+    service.insert(Collections.singletonList(createKafkaRecord(simpleRecordJson, 2, false)));
+
+    rows = describeTable(tableName);
+    assertThat(rows.size()).isEqualTo(10);
+    assertThat(rows).contains(new DescribeTableRow("SIMPLE", "VARCHAR(16777216)"));
+
+    // reinsert record with extra field
+    service.insert(Collections.singletonList(createKafkaRecord(simpleRecordJson, 2, false)));
+    waitForOffset(3);
+  }
+
+  private static Stream<Arguments> prepareData() {
+    return Stream.of(
+        Arguments.of(
+            "Primitive JSON with schema",
+            primitiveJsonWithSchema,
+            new DescribeTableRow[] {
+              new DescribeTableRow("ID_INT8", "NUMBER(10,0)"),
+              new DescribeTableRow("ID_INT16", "NUMBER(10,0)"),
+              new DescribeTableRow("ID_INT32", "NUMBER(10,0)"),
+              new DescribeTableRow("ID_INT64", "NUMBER(19,0)"),
+              new DescribeTableRow("DESCRIPTION", "VARCHAR(16777216)"),
+              new DescribeTableRow("RATING_FLOAT32", "FLOAT"),
+              new DescribeTableRow("RATING_FLOAT64", "FLOAT"),
+              new DescribeTableRow("APPROVAL", "BOOLEAN")
+            },
+            true),
+        Arguments.of(
+            "Primitive JSON without schema",
+            primitiveJson,
+            new DescribeTableRow[] {
+              new DescribeTableRow("ID_INT8", "NUMBER(19,0)"),
+              new DescribeTableRow("ID_INT16", "NUMBER(19,0)"),
+              new DescribeTableRow("ID_INT32", "NUMBER(19,0)"),
+              new DescribeTableRow("ID_INT64", "NUMBER(19,0)"),
+              new DescribeTableRow("DESCRIPTION", "VARCHAR(16777216)"),
+              new DescribeTableRow("RATING_FLOAT32", "FLOAT"),
+              new DescribeTableRow("RATING_FLOAT64", "FLOAT"),
+              new DescribeTableRow("APPROVAL", "BOOLEAN")
+            },
+            false));
+  }
+}
